mostly pep8 compliant

+++ scripts/zim/dedupzim.py
@@ -8,19 +8,22 @@ import tempfile
 
 DIVIDER = 'Tasks Completed'
 
+
 class TaskItem(object):
     """A class for tasks that's nestable."""
     re_detab = re.compile(r'^\t')
-    def __init__(self, tasks_paragraph):
-        tasks = tasks_paragraph.split('\n')
+
+    def __init__(self, tasks):
         self.task = tasks[0].rstrip()
         if len(tasks) > 1:
             paragraphs = self.make_paragraphs(self.re_detab.sub('', i) for i in tasks[1:])
-            self.subtasks = set(TaskItem(i) for i in paragraphs if i != '')
+            self.subtasks = set(TaskItem(i) for i in paragraphs if i != [''])
         else:
             self.subtasks = set()
+
     def join(self, other):
-        if self.task != other.task: raise ValueError
+        if self.task != other.task:
+            raise ValueError
         newTask = TaskItem(self.task)
         newTask.subtasks = self.subtasks.symmetric_difference(other.subtasks)
         intersect = self.subtasks.intersection(other.subtasks)
@@ -32,14 +35,19 @@ class TaskItem(object):
                     j = dict_you[k]
                     newTask.subtasks.add(v.join(j))
         return newTask
+
     def display(self):
         return "\n".join([self.task] + ['\t' + i.display() for i in self.subtasks])
+
     def __repr__(self):
         return self.task
+
     def __lt__(self, other):
         return self.task < other.task
+
     def __eq__(self, other):
         return self.task == other.task
+
     def __hash__(self):
         return hash(self.task)
 
@@ -50,20 +58,21 @@ class TaskItem(object):
             if line.strip() == '':
                 continue
             elif '\t' in line:
-                out[-1] += line
+                out[-1].append(line)
             else:
-                out.append(line)
-        return (i for i in out)
+                out.append([line])
+        return (tuple(i) for i in out)
 
-def make_tasks(paragraphs, divider = DIVIDER):
+
+def make_tasks(paragraphs, divider=DIVIDER):
     tasks = set()
     done = set()
     out = []
 
     for line in paragraphs:
-        if '' == line.strip():
+        if line == ['']:
             continue
-        elif '[ ]' in line:
+        elif any('[ ]' in i for i in line):
             task = TaskItem(line)
             if task not in tasks:
                 out.append(task)
@@ -71,31 +80,40 @@ def make_tasks(paragraphs, divider = DIVIDER):
             else:
                 n = out.index(task)
                 out[n] = output[n].join(task)
-        elif '[*]' in line:
+        elif any('[*]' in i for i in line):
             done.add(line)
-        elif divider in line:
+        elif any(divider in i for i in line):
             break
-        else: out.append(line.rstrip())
+        else:
+            out.append([i.rstrip() for i in line])
 
-    out.append('==== %s ====\n' % divider)
+    out.append(['==== %s ====\n' % divider])
     out += list(done)
-    for line in paragraphs: out.append(line)
+    for line in paragraphs:
+        out.append(line)
     return out
 
-def display_tasks(out, outfile = sys.stdout):
+
+def display_tasks(out, outfile=sys.stdout):
     for i in out:
         if isinstance(i, TaskItem):
             outfile.write(i.display() + '\n\n')
         else:
-            outfile.write(i + '\n')
+            outfile.write("\n".join(i) + '\n')
+
 
 def open_infile(infile_name=None):
+    """Open input file. Reads from stdin if no
+    input file"""
     if infile_name is None:
         return sys.stdin
     else:
         return open(infile_name)
 
-def open_outfile(infile_name=None, outfile_name=None, in_place = False):
+
+def open_outfile(infile_name=None, outfile_name=None, in_place=False):
+    """Open output file. If in_place or output file same as input file,
+    open tempfile instead. Elif no outfile, output to stdout."""
     if in_place or infile_name == outfile_name:
         outfile = tempfile.NamedTemporaryFile(mode='w+t', delete=False)
         path = outfile.name
@@ -106,30 +124,32 @@ def open_outfile(infile_name=None, outfile_name=None, in_place = False):
         outfile = open(outfile_name)
     return outfile, None
 
-def main(infile_name = None, outfile_name = None, in_place = False, divider = DIVIDER):
+
+def main(infile_name=None, outfile_name=None, in_place=False, divider=DIVIDER):
     try:
         infile = open_infile(infile_name)
         paragraphs = TaskItem.make_paragraphs(infile)
     finally:
-        if infile != sys.stdin:
+        if "infile" in locals() and infile != sys.stdin:
             infile.close()
     out = make_tasks(paragraphs, divider)
     try:
         outfile, tmppath = open_outfile(infile_name, outfile_name, in_place)
         display_tasks(out, outfile)
     finally:
-        if outfile != sys.stdout:
+        if "outfile" in locals() and outfile != sys.stdout:
             outfile.close()
-    if tmppath:
+            os.remove(tmppath)
+    if "tmppath" in locals() and tmppath:
         shutil.copy(tmppath, infile_name)
         os.remove(tmppath)
+
 
 if __name__ == '__main__':
     import argparse
     parser = argparse.ArgumentParser()
     parser.add_argument('infile', nargs='?', type=str, default=None)
     parser.add_argument('outfile', nargs='?', type=str, default=None)
-    #in-place option? Switch infile and outfile into strings so I can validate and then open files myself? Something with tempfiles?
     parser.add_argument('-i', '--in-place', help='Output file. Default is stdout', action="store_true")
     parser.add_argument('-d', '--divider', default=DIVIDER)
     args = parser.parse_args()
